pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "creditscope"
version = "1.0.0"
description = "Agentic Credit Scoring with MoE Observability"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Sarel Weinberger", email = "sarelw1@gmail.com" }
]
keywords = ["credit-scoring", "llm", "moe", "observability", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
backend = [
    "Pillow>=10.4.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.12",
]
inference = [
    "torch>=2.4.0",
    "sglang[all]>=0.4.5",
    "transformers>=4.46.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
all = [
    "creditscope[backend]",
    "creditscope[inference]",
    "creditscope[dev]",
]

[project.urls]
Homepage = "https://github.com/sarelWeinberger/creditscope"
Documentation = "https://github.com/sarelWeinberger/creditscope#readme"
Repository = "https://github.com/sarelWeinberger/creditscope.git"
Issues = "https://github.com/sarelWeinberger/creditscope/issues"

[tool.hatch.build.targets.wheel]
packages = ["backend", "inference"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["backend", "inference"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sglang.*",
    "pytesseract.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=backend --cov=inference --cov-report=term-missing"
